null

+++ src/main/java/com/http/shiller/first/vehicles/base/util/VehicleUtil.java
@@ -3,7 +3,7 @@ package com.http.shiller.first.vehicles.base.util;
 import com.http.shiller.first.vehicles.Vehicle;
 
 public class VehicleUtil {
-    public static int calculatePrice(Vehicle vehicle) {
+    public static double calculatePrice(Vehicle vehicle) {
         int basePrice = vehicle.getBasePrice();
         double yearCoeff = 1 + ((vehicle.getProducedYear() - 1950) * 0.01);
         double milesCoeff = 1;
@@ -13,7 +13,7 @@ public class VehicleUtil {
         if (vehicle.getMiles() > 10000) {
             milesCoeff = 0.7;
         }
-        return (int)(basePrice * yearCoeff * milesCoeff);
+        return basePrice * yearCoeff * milesCoeff;
     }
 
     public static Vehicle compare(Vehicle vehicle1, Vehicle vehicle2) {
